salud/requirements.txt
appdirs==1.4.3
coreapi==2.3.0
coreschema==0.0.4
Django==1.11.11
django-cors-headers==2.0.2
django-filter==1.0.4
django-rest-auth==0.9.1
django-rest-framework-docs==0.1.7
django-rest-swagger==2.1.2
djangorestframework==3.6.3
drf-any-permissions==0.0.1
dry-rest-permissions==0.1.9
itypes==1.1.0
Jinja2==2.9.6
jsonpickle==0.9.4
Markdown==2.6.8
MarkupSafe==1.0
mysqlclient==1.3.10
openapi-codec==1.3.1
packaging==16.8
Pillow==5.0.0
wheel==0.31.1
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.4
pyparsing==2.2.0
pytz==2018.3
requests==2.14.2
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
